pyproject.toml
[project]
name = "ob"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pygithub>=2.8.1",
    "typer>=0.20.0",
]

[project.scripts]
ob = "ob.cli:app"

[build-system]
requires = ["uv_build >= 0.8.17, <0.9.0"]
build-backend = "uv_build"
